requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
apscheduler==3.10.4
firebase-admin==6.5.0
python-dotenv==1.0.1
pytz==2024.1

# Dev / test
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
